Split sprite into world region and sheet region.
This is nice because we don't have spritesheet data competing for
cache space with location data, so we can efficiently copy world
position data into the sprite renderer in one memcpy (or use the
world's position data as the storage for object positions).

Test2d also now has some examples of timing control.