draft of automatic upload tracking for frenderer.
it's kind of complicated code which I don't love, and it mainly --
what -- saves you from having to remember to call upload_sprites or
whatever?  While it's true that can be a real confusing failure mode
is it worth the extra code?

This shouldn't impact the ability to use frenderer's renderers outside
of frenderer.